Allow task names to be set before creating them
Instead of forcing users to set the name of a new task by editing the
name of a newly created task, support setting the task name *before*
actually creating the task.

This allows us to remove the code that, as a convenience for the user,
simulated a click on each newly created task so that the task name
could be updated easily, which makes one item on the TODO list in the
README go away.